requirements.txt
fastapi
uvicorn
python-multipart
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
